requirements.txt
arrow==0.12.1
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
BareNecessities==0.2.8
beautifulsoup4==4.6.0
blinker==1.4
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.6
constantly==15.1.0
cryptography==2.3
cycler==0.10.0
dialogflow==0.5.1
dj-database-url==0.5.0
Django==2.1
django-heroku==0.3.1
docopt==0.6.2
eloquent==0.5
enum-compat==0.0.2
eventlet==0.21.0
Flask==1.0.2
flask-mail==0.9.1
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
Flask-SocketIO==2.9.2
Flask-SQLAlchemy==2.3.2
google-api-core==1.3.0
google-auth==1.5.1
googleapis-common-protos==1.5.3
greenlet==0.4.14
grpcio==1.14.2
gunicorn==19.6.0
hdfs==2.1.0
heroku==0.1.4
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
inflection==0.3.1
inline==0.0.1
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
Mail==2.1.0
MarkupSafe==0.23
matplotlib==2.2.2
mysqlclient==1.3.13
ndg-httpsclient==0.5.1
ngrok==0.0.1
node==0.9.22
npm==0.1.1
numpy==1.14.5
odict==1.6.2
optional-django==0.1.0
pandas==0.23.3
passlib==1.7.1
php==1.2.0
pipenv==2018.7.1
plumber==1.5
protobuf==3.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pusher==2.0.1
py4j==0.10.7
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyHamcrest==1.9.0
pymessenger==0.0.7.0
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pyparsing==2.2.0
pypiwin32==223
pyspark==2.3.1
python-dateutil==1.5
python-dotenv==0.9.1
python-engineio==1.7.0
python-socketio==1.8.1
pytz==2018.5
pywapi==0.3.8
pywin32==223
requests==2.19.1
requests-toolbelt==0.8.0
rsa==3.4.2
scikit-learn==0.19.2
simplejson==3.16.0
six==1.11.0
SQLAlchemy==1.2.10
Twisted==18.7.0
unicode==2.6
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper-win==1.2.5
Werkzeug==0.14.1
whitenoise==3.3.1
WTForms==2.2.1
yahooweather==0.10
zope.component==4.4.1
zope.deprecation==4.3.0
zope.event==4.3.0
zope.interface==4.5.0
zope.lifecycleevent==4.2.0